refactor: 경로 및 파일 이름 수정 #141
