+++ src/app/concert/form/[id]/view/page.tsx
@@ -6,8 +6,8 @@ import { useRouter } from 'next/navigation';
 import FormInfo from '@/app/concert/form/[id]/_shared/components/form-info/form-info';
 import FormCancelModal from '@/app/concert/form/[id]/_shared/components/form-modal/form-cancel-modal';
 import FormConfirmModal from '@/app/concert/form/[id]/_shared/components/form-modal/form-confirm-modal';
-import FormReasonModal from '@/app/concert/form/[id]/_shared/components/form-modal/form-reason-modal';
-import FormRejectedModal from '@/app/concert/form/[id]/_shared/components/form-modal/form-rejected-modal';
+import FormReasonModal from '@/app/concert/form/[id]/_shared/components/form-modal/form-reason-modal/form-reason-modal';
+import FormRejectedModal from '@/app/concert/form/[id]/_shared/components/form-modal/form-rejected-modal/form-rejected-modal';
 import FormTabManager from '@/app/concert/form/[id]/_shared/components/form-tab/form-tab-manager';
 import { useGetFormDetail } from '@/app/concert/form/[id]/_shared/services/query';
 import PageFrame from '@/shared/components/layout/page-frame/page-frame';
